requirements.txt
adafruit-io
pyserial
pyyaml
